Version du 26/03/2020

+++ src/ch/epfl/rigel/math/Angle.java
@@ -34,8 +34,7 @@ public final class Angle {
     }
 
     /**
-     * Returns the angle corresponding to the given number of arc seconds, in
-     * radians.
+     * Returns the angle corresponding to the given number of arc seconds, in radians.
      *
      * @param sec
      *            The number of arc seconds
@@ -46,8 +45,7 @@ public final class Angle {
     }
 
     /**
-     * Returns the angle corresponding to the given angle deg​° min​′ sec​″, in
-     * radians.
+     * Returns the angle corresponding to the given angle deg​° min​′ sec​″, in radians.
      *
      * @param deg
      *            The given number of degrees
@@ -55,22 +53,19 @@ public final class Angle {
      *            The given number of minutes
      * @param sec
      *            The given number of seconds
-     * @return the angle in radians
-     *
      * @throws IllegalArgumentException
-     *             if the interval [0,60[ does not contain the number of minutes
-     *             or the number of seconds
+     *             if the interval [0,60[ does not contain the given number of minutes and/or seconds
+     * @return the angle in radians
      */
     public static double ofDMS(int deg, int min, double sec) {
         double correctMin = Preconditions.checkInInterval(DMS_INTERVAL, min);
         double correctSec = Preconditions.checkInInterval(DMS_INTERVAL, sec);
 
-        return Math.toRadians((double) deg + (double) correctMin / 60 + correctSec / 3600);
+        return Math.toRadians((double) deg + correctMin / 60 + correctSec / 3600);
     }
 
     /**
-     * Returns the angle corresponding to the given angle in degrees, in
-     * radians.
+     * Returns the angle corresponding to the given angle in degrees, in radians.
      *
      * @param deg
      *            The given angle in degrees
@@ -81,8 +76,7 @@ public final class Angle {
     }
 
     /**
-     * Returns the angle corresponding to the given angle in radians, in
-     * degrees.
+     * Returns the angle corresponding to the given angle in radians, in degrees.
      *
      * @param rad
      *            The given angle in radians
@@ -113,5 +107,4 @@ public final class Angle {
     public static double toHr(double rad) {
         return rad * HOUR_PER_RAD;
     }
-
 }
